move resolveScheduleToDate to its own file
