+++ mod.ts
@@ -73,7 +73,8 @@ export type {
   ApiKeyCredential,
   CredentialFetcher,
 } from "./runtimeSupport.ts";
-export type { DelayedTask, DelayedTaskSchedule } from "./tasks.ts";
+export type { DelayedTask } from "./tasks.ts";
+export type { DelayedTaskSchedule, DelayedTaskTimePeriod } from "./tasks/schedule.ts";
 export type { SecretFetcher, SecretFetcherOptions } from "./secrets.ts";
 export type { CommonCredentialFetcherOptions, CommonTriggerOptions } from "./common.ts";
